+++ src/modules/feature/room/RoomBoardGridItem.vue
