+++ src/modules/feature/room/index.ts
@@ -1,7 +1,7 @@
 import RoomAdminMembersTable from "./manageRoom/tables/RoomAdminMembersTable.vue";
 import RoomAdminTable from "./manageRoom/tables/RoomAdminTable.vue";
 import Registration from "./registration/Registration.vue";
-import RoomContentGrid from "./RoomContentGrid.vue";
+import RoomBoardGrid from "./RoomBoardGrid.vue";
 import RoomForm from "./RoomForm.vue";
 import RoomGrid from "./RoomGrid.vue";
 import AddExternalPersonDialog from "./roomMembers/dialogs/add-external-person-dialog/AddExternalPersonDialog.vue";
@@ -31,7 +31,7 @@ export {
 	Registration,
 	RoomAdminMembersTable,
 	RoomAdminTable,
-	RoomContentGrid,
+	RoomBoardGrid,
 	RoomForm,
 	RoomGrid,
 	RoomMenu,
